clearbill/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
anthropic
pandas
numpy
httpx
websockets
python-dotenv
